Put collect_events content in map_events
