+++ blocks/antelope/src/lib.rs
@@ -4,11 +4,10 @@ mod auth_sequences;
 mod authority;
 mod authorizations;
 mod blocks;
-mod collect_events;
 mod creation_tree;
 mod db_ops;
+mod events;
 mod feature_ops;
-mod map_events;
 mod pb;
 mod perm_ops;
 mod ram_ops;
